backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
termcolor==3.3.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.0
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
wheel==0.46.3
wrapt==2.1.2
groq
python-dotenv
slowapi==0.1.9
